Add aeds/kvs package for key-value convenience
It's often convenient to store small amounts of global state in a
key-value database.  In many circumstances, memcache alone is not good
enough because we want more predictable guarantees about data
lifetime (can't afford data to disappear when caches get full).

The kvs package gives a simple key-value API and takes care of
persisting data in memcache and datastore along with automatic
expiration.