Write to memcache after field mismatch read
If reading an entity from the datastore succeeds but encountered a field
mismatch error, we treat it the same as if it succeeded without any
error at all.  The old code failed to write to memcache in this
circumastance which cause us to read from datastore more often than
necessary.