Put: delete from, don't write to, memcache
When writing an entity with aeds.Put() we want memcache to portray a
consistent view of the data.  We used to try for this consistency by
writing to memcache after the datastore write succeeded.  However, the
best practice appears to be that we should simply delete the memcache
entry and allow a future read to repopulate it.

This increases datastore read costs slightly but should give better
consistency.  By forcing the first read after a write to hit the
datastore, we should also get better interoperability with transactions
since the datastore knows the read is taking place.

See https://cloud.google.com/appengine/docs/python/ndb/cache#memcache