Add Modify() for atomic read-modify-write
Performing a read-modify-write inside a datastore transaction requires
quite a bit of boilerplate code.  When using memcache, that codes get
more complicated and one is likely to get it wrong.

Modify() hides all the boilerplate and complexity.  One's code need only
modify the contents of a struct and all the details are taken care of.