CollectGarbage: more flexible API
Different applications will generate KV garbage at much different rates.
We're going to want knobs for adjusting the details.  That's easier to
do with named arguments for which we provide sensible defaults.

While I was changing the API, I also added the number of entities
removed.  This can be a helpful metric for knowing when the GC knobs
need adjusting.